fix: AttributeError when no EXPIRING_TOKEN_LIFESPAN in settings.py file
 - let the token generation and authentication to happen without expiration time, like
default DRF token authentication, if no EXPIRING_TOKEN_LIFESPAN is declared in settings file

+++ rest_framework_expiring_authtoken/views.py
@@ -26,6 +26,10 @@ class ObtainExpiringAuthToken(ObtainAuthToken):
                 user=serializer.validated_data['user']
             )
 
+            if not token.expiration_set():
+                # If no EXPIRING_TOKEN_LIFESPAN is set in settings file, then skip checking if token is expired
+                return Response({'token': token.key})
+
             if token.expired():
                 # If the token is expired, generate a new one.
                 token.delete()
